# src/docs/swagger.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "tp-devops",
		"description": "Trabajo practico devops",
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:3000"
		},
		{
			"url": "https://tp-devops-o6jb.onrender.com"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"responses": {
			"Error": {
				"description": "Failed",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"message": {
									"type": "string"
								},
								"error": {
									"type": "object",
									"properties": {
										"status": {
											"type": "integer"
										}
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"paths": {
		"/": {
			"get": {
				"operationId": "root",
				"description": "Root",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/html": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"5XX": {
						"description": "Unexpected error"
					}
				}
			}
		},
		"/ping": {
			"get": {
				"operationId": "ping",
				"description": "Check service version",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"version": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"5XX": {
						"description": "Unexpected error"
					}
				}
			}
		},
		"/health": {
			"get": {
				"operationId": "health",
				"description": "Check weather service is working",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"5XX": {
						"description": "Unexpected error"
					}
				}
			}
		},
		"/error": {
			"get": {
				"operationId": "error",
				"description": "Return an error",
				"responses": {
					"500": {
						"$ref": "#/components/responses/Error"
					}
				}
			}
		},
		"/users": {
			"get": {
				"security": [
					{
						"bearerAuth": []
					}
				],
				"operationId": "get-users",
				"description": "Get users",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"users": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"name": {
														"type": "string"
													},
													"email": {
														"type": "string"
													},
													"role": {
														"type": "string"
													},
													"active": {
														"type": "boolean"
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"5XX": {
						"description": "Unexpected error"
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"security": [
					{
						"bearerAuth": []
					}
				],
				"operationId": "get-user",
				"description": "Get user by id",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "user id",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 24,
							"maxLength": 24
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"user": {
											"type": "object",
											"properties": {
												"name": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"role": {
													"type": "string"
												},
												"active": {
													"type": "boolean"
												}
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Error"
					},
					"404": {
						"$ref": "#/components/responses/Error"
					},
					"5XX": {
						"description": "Unexpected error"
					}
				}
			}
		},
		"/login": {
			"post": {
				"operationId": "login",
				"description": "Login",
				"requestBody": {
					"description": "login",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"username": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Login successfull",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"token": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Error"
					},
					"422": {
						"$ref": "#/components/responses/Error"
					},
					"500": {
						"$ref": "#/components/responses/Error"
					}
				}
			}
		}
	}
}